+++ game/components/c_timer.h
@@ -1,0 +1,15 @@
+//project verse, 2017-2021
+//by jose pazos perez
+//all rights reserved uwu
+
+#pragma once
+
+#include "dtypes.h"
+
+namespace Verse::Component
+{
+    struct Timer {
+        ui32 tid = 0;
+        ui16 ms;
+    };
+}