+++ game/controllers/controller_list.h
@@ -19,5 +19,7 @@
 #define CAMERA_FREE_CONTROLLER [&c, eid]() -> bool {return Controller::Camera::Free::controller(c, eid);};
 
 #include "moving_platform_controller.h"
+#include "falling_platform_controller.h"
 
 #define MOVING_PLATFORM_CONTROLLER [&c, eid, move]() -> bool {return Controller::MovingPlatform::controller(c, eid, move);};
+#define FALLING_PLATFORM_CONTROLLER [&c, eid, move]() -> bool {return Controller::FallingPlatform::controller(c, eid, move);};